fix(test): Fix missing primary key test
