+++ etl/tests/integration/mod.rs
@@ -1,4 +1,3 @@
-mod no_primary_key_test;
 mod pipeline_test;
 mod postgres_store;
 mod replica_identity;
