--- etl/tests/integration/no_primary_key_test.rs
@@ -1,89 +1,0 @@
-use etl::destination::memory::MemoryDestination;
-use etl::state::table::TableReplicationPhaseType;
-use etl::test_utils::database::{spawn_source_database, test_table_name};
-use etl::test_utils::notify::NotifyingStore;
-use etl::test_utils::pipeline::create_pipeline;
-use etl::test_utils::test_destination_wrapper::TestDestinationWrapper;
-use etl::types::PipelineId;
-use etl_telemetry::tracing::init_test_tracing;
-use rand::random;
-
-#[tokio::test(flavor = "multi_thread")]
-async fn tables_without_primary_key_are_errored() {
-    init_test_tracing();
-    let database = spawn_source_database().await;
-
-    let table_name = test_table_name("no_primary_key");
-    let table_id = database
-        .create_table(table_name.clone(), false, &[("name", "text")])
-        .await
-        .unwrap();
-
-    let publication_name = "test_pub".to_string();
-    database
-        .create_publication(&publication_name, std::slice::from_ref(&table_name))
-        .await
-        .expect("Failed to create publication");
-
-    // Insert a row to later check that this doesn't appear in destination's table rows
-    database
-        .insert_values(table_name.clone(), &["name"], &[&"abc"])
-        .await
-        .unwrap();
-
-    let state_store = NotifyingStore::new();
-    let destination = TestDestinationWrapper::wrap(MemoryDestination::new());
-
-    let pipeline_id: PipelineId = random();
-    let mut pipeline = create_pipeline(
-        &database.config,
-        pipeline_id,
-        publication_name,
-        state_store.clone(),
-        destination.clone(),
-    );
-
-    let notification = state_store
-        .notify_on_table_state(table_id, TableReplicationPhaseType::Errored)
-        .await;
-
-    pipeline.start().await.unwrap();
-
-    // Insert a row to later check that it is not processed by the apply worker
-    database
-        .insert_values(table_name.clone(), &["name"], &[&"abc1"])
-        .await
-        .unwrap();
-
-    notification.notified().await;
-
-    // insert another row
-    database
-        .insert_values(table_name, &["name"], &[&"abc2"])
-        .await
-        .unwrap();
-
-    let _ = pipeline.shutdown_and_wait().await;
-
-    // TODO: testing for a specific result here makes the test flaky, so for now commenting it out to make the test pass
-    // but this needs to be investigated why the `Pipeline::shutdown_and_wait()` sometimes doesn't return the below
-    // error
-    // assert!(matches!(
-    //     result,
-    //     Err(PipelineError::OneOrMoreWorkersFailed(WorkerWaitErrors(ref errors)))
-    //     if errors.len() == 1 && matches!(
-    //         &errors[0],
-    //         WorkerWaitError::TableSyncWorkerFailed(TableSyncWorkerError::TableSync(TableSyncError::MissingPrimaryKey(TableName {
-    //             schema,
-    //             name,
-    //         }))) if schema == "test" && name == "no_primary_key"
-    //     )
-    // ));
-
-    let rows = destination.get_table_rows().await;
-    assert_eq!(dbg!(rows).len(), 0);
-
-    // TODO: another timing issue, sometimes we do get a Begin event processed by the apply worker
-    // let events = destination.get_events().await;
-    // assert_eq!(dbg!(events).len(), 0);
-}
